made inner class RandomPositionGenerator a top level class
