+++ gazeplay/src/main/java/net/gazeplay/commons/utils/Portrait.java
@@ -1,14 +1,10 @@
 package net.gazeplay.commons.utils;
 
-import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.scene.paint.ImagePattern;
 import javafx.scene.shape.Circle;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import net.gazeplay.commons.utils.games.Utils;
-
-import java.util.Random;
 
 /**
  * Created by schwab on 12/08/2016.
@@ -19,15 +15,12 @@ public class Portrait extends Circle {
         return Utils.images(Utils.getImagesFolder() + "portraits");
     }
 
-    private final RandomPositionGenerator randomPositionGenerator;
-
     @Getter
     private final int initialRadius;
 
     public Portrait(int initialRadius, RandomPositionGenerator randomPositionGenerator, Image[] availableImages) {
         super(initialRadius);
         this.initialRadius = initialRadius;
-        this.randomPositionGenerator = randomPositionGenerator;
 
         this.setPosition(randomPositionGenerator.newRandomPosition(initialRadius));
 
@@ -51,30 +44,6 @@ public class Portrait extends Circle {
         int count = availableImages.length;
         int index = (int) (count * Math.random());
         return availableImages[index];
-    }
-
-    @AllArgsConstructor
-    public static class RandomPositionGenerator {
-
-        private final Random random = new Random();
-
-        private final Scene scene;
-
-        public Position newRandomPosition(double radius) {
-
-            // if (scene.getWidth() == 0 || scene.getHeight() == 0) {
-            // return new Position(0, 0);
-            // }
-
-            double minX = radius;
-            double minY = radius;
-            double maxX = scene.getWidth() - radius;
-            double maxY = scene.getHeight() - radius;
-            double positionX = random.nextInt((int) (maxX - minX)) + minX;
-            double positionY = random.nextInt((int) (maxY - minY)) + minY;
-            return new Position((int) positionX, (int) positionY);
-        }
-
     }
 
 }
