+++ gazeplay/src/main/java/net/gazeplay/games/creampie/Target.java
@@ -15,6 +15,7 @@ import net.gazeplay.commons.gaze.GazeEvent;
 import net.gazeplay.commons.gaze.GazeUtils;
 import net.gazeplay.commons.utils.Portrait;
 import net.gazeplay.commons.utils.Position;
+import net.gazeplay.commons.utils.RandomPositionGenerator;
 import net.gazeplay.commons.utils.stats.ShootGamesStats;
 import net.gazeplay.commons.utils.stats.Stats;
