+++ gazeplay/src/main/java/net/gazeplay/games/ninja/Ninja.java
@@ -4,6 +4,7 @@ import javafx.scene.Scene;
 import net.gazeplay.GameContext;
 import net.gazeplay.GameLifeCycle;
 import net.gazeplay.commons.utils.Portrait;
+import net.gazeplay.commons.utils.RandomPositionGenerator;
 import net.gazeplay.commons.utils.stats.ShootGamesStats;
 
 /**
@@ -25,7 +26,7 @@ public class Ninja implements GameLifeCycle {
     public void launch() {
         Scene scene = gameContext.getScene();
 
-        Portrait.RandomPositionGenerator randomPositionGenerator = new Portrait.RandomPositionGenerator(scene);
+        RandomPositionGenerator randomPositionGenerator = new RandomPositionGenerator(scene);
 
         Target portrait = new Target(gameContext, randomPositionGenerator, stats, Portrait.loadAllImages());
