+++ gazeplay/src/main/java/net/gazeplay/games/ninja/Target.java
@@ -14,6 +14,7 @@ import net.gazeplay.commons.gaze.GazeEvent;
 import net.gazeplay.commons.gaze.GazeUtils;
 import net.gazeplay.commons.utils.Portrait;
 import net.gazeplay.commons.utils.Position;
+import net.gazeplay.commons.utils.RandomPositionGenerator;
 import net.gazeplay.commons.utils.stats.Stats;
 
 import java.net.URL;
